requirements.txt
streamlit==1.52.1
tensorflow==2.20.0
numpy
opencv-python
pillow
